remove ui file and abstract ontap function
