+++ packages/smooth_app/lib/pages/product/attribute_first_row_widget.dart
@@ -1,15 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:smooth_app/pages/product/atrribute_first_row_helper.dart';
+import 'package:openfoodfacts/openfoodfacts.dart';
+import 'package:smooth_app/pages/product/attribute_first_row_helper.dart';
 
 class AttributeFirstRowWidget extends StatefulWidget {
-  const AttributeFirstRowWidget({
-    required this.helper,
-    required this.onTap,
-  });
+  const AttributeFirstRowWidget({required this.helper, required this.product});
 
   final AttributeFirstRowHelper helper;
-  final Function() onTap;
+  final Product product;
 
   @override
   State<AttributeFirstRowWidget> createState() =>
@@ -60,7 +58,11 @@ class _AttributeFirstRowWidgetState extends State<AttributeFirstRowWidget> {
           ),
           iconColor: theme.primaryColor,
           tileColor: theme.colorScheme.secondary,
-          onTap: widget.onTap,
+          onTap: widget.helper.onTap(
+            context: context,
+            product: widget.product,
+            mounted: mounted,
+          ),
         ),
         _termsList(
           showAllTerms ? allTerms : firstTerms,
